requirements.txt
Django==5.2
djangorestframework==3.16.0
pillow==11.1.0